requirements.txt
Django==4.2.4
isort==5.12.0
black==23.7.0
flake8==6.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
markdown==3.5.1
django-filter==23.5 
djangorestframework-simplejwt==5.3.1